+++ Client.Test/PointDataTest.cs
@@ -34,6 +34,16 @@ namespace InfluxDB.Client.Test
         }
         
         [Test]
+        public void EqualSignEscaping() {
+
+            var point = PointData.Measurement("h=2o")
+                .Tag("l=ocation", "e=urope")
+                .Field("l=evel", 2);
+
+            Assert.AreEqual("h=2o,l\\=ocation=e\\=urope l\\=evel=2i", point.ToLineProtocol());
+        }
+        
+        [Test]
         public void Immutability()
         {
             var point = PointData.Measurement("h2 o")
@@ -72,7 +82,7 @@ namespace InfluxDB.Client.Test
                 .Tag("", "warn")
                 .Field("level", 2);
 
-            Assert.AreEqual("h2\\=o,location=europe level=2i", point.ToLineProtocol());
+            Assert.AreEqual("h2=o,location=europe level=2i", point.ToLineProtocol());
 
             point = PointData.Measurement("h2,o")
                 .Tag("location", "europe")